requirements.txt
aiofiles==23.1.0
aiologger==0.7.0
aioschedule==0.5.2
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
et-xmlfile==1.1.0
Flask
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
openpyxl==3.1.2
paramiko==3.1.0
pycparser==2.21
pydantic==1.10.7
PyNaCl==1.5.0
PySocks==1.7.1
pyusb==1.2.1
pyvmlib==2.4.1
pyvmomi==8.0.1.0
PyYAML==6.0
requests==2.30.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
typing-extensions==4.5.0
urllib3==2.0.2
zipp==3.15.0
